requirements.txt
anaconda-client==1.2.2
bcrypt==3.1.7
blinker==1.4
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
clyent==1.2.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==2.2.1
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
